Check for primary key in insertData

+++ table.h
@@ -44,10 +44,10 @@ class Table{
 		bool isDropped();
 		std::string getFormattedMetaData();
 		std::string parseRecord(char* buffer);
-		int getColumnByteSizeAt(int columnIndex);
-		int getColumnBytePosition(std::string clmn);
-		int getColumnByteSize(std::string);
-		int getIndexOfPrimaryKey();
+		size_t getColumnByteSizeAt(int columnIndex);
+		size_t getColumnBytePosition(std::string clmn);
+		size_t getColumnByteSize(std::string);
+		size_t getIndexOfPrimaryKey();
 };
 
 #endif
